Rename entities to the proper naming
